Move SignedStateSentinel
