+++ platform-sdk/consensus-roster/src/main/java/module-info.java
@@ -5,11 +5,11 @@ module org.hiero.consensus.roster {
 
     requires transitive com.hedera.node.hapi;
     requires transitive com.hedera.pbj.runtime;
+    requires transitive com.swirlds.metrics.api;
     requires transitive com.swirlds.state.api;
     requires transitive org.hiero.base.crypto;
     requires transitive org.hiero.consensus.model;
     requires com.swirlds.base;
-    requires transitive com.swirlds.metrics.api;
     requires org.hiero.base.utility;
     requires org.hiero.consensus.metrics;
     requires static transitive com.github.spotbugs.annotations;
