+++ platform-sdk/consensus-state-management/src/main/java/org/hiero/consensus/state/management/config/StateManagementWiringConfig.java
@@ -4,6 +4,7 @@ package org.hiero.consensus.state.management.config;
 import com.swirlds.component.framework.schedulers.builders.TaskSchedulerConfiguration;
 import com.swirlds.config.api.ConfigData;
 import com.swirlds.config.api.ConfigProperty;
+import java.time.Duration;
 
 /**
  * Contains configuration values for the state management wiring.
@@ -13,6 +14,8 @@ import com.swirlds.config.api.ConfigProperty;
  * @param stateSigner configuration for the state signer scheduler
  * @param stateSignatureCollector configuration for the state signature collector scheduler
  * @param stateSnapshotManager configuration for the state snapshot manager scheduler
+ * @param signedStateSentinel configuration for the signed state sentinel scheduler
+ * @param signedStateSentinelHeartbeatPeriod the frequency of signed state sentinels heartbeats
  */
 @ConfigData("state.management.wiring")
 public record StateManagementWiringConfig(
@@ -31,4 +34,9 @@ public record StateManagementWiringConfig(
         TaskSchedulerConfiguration stateSignatureCollector,
 
         @ConfigProperty(defaultValue = "SEQUENTIAL_THREAD CAPACITY(20) UNHANDLED_TASK_METRIC")
-        TaskSchedulerConfiguration stateSnapshotManager) {}
+        TaskSchedulerConfiguration stateSnapshotManager,
+
+        @ConfigProperty(defaultValue = "SEQUENTIAL UNHANDLED_TASK_METRIC")
+        TaskSchedulerConfiguration signedStateSentinel,
+
+        @ConfigProperty(defaultValue = "10s") Duration signedStateSentinelHeartbeatPeriod) {}
